refactor runs filters
