+++ src/renderer/components/Runs.vue
@@ -1,11 +1,6 @@
 <template>
     <section class="section runs">
-        <div v-if="allRuns && allRuns.length > 0" class="filters">
-            <div class="search">
-                <input v-model="filterText" @input="resetPagination" placeholder="Search runs">
-            </div>
-            <CustomSelect v-if="allTags && tagsWithRuns.length > 0" type="multi" :items="tagsWithRuns" label="Tags" emptyMessage="No tags selected" @updateSelect="onUpdateTagsMultiSelect"/>
-        </div>
+        <RunsFilters :filter-offset="filterOffset" :filter-limit-per-page="filterLimitPerPage" :filter-order="filterOrder" @filteredRunsTotal="onUpdateFilteredRunsTotal" @filteredRuns="onUpdateFilteredRuns" @resetPagination="resetPagination"/>
         <transition-group name="run-group-transition" tag="ul" class="runs-container">
             <template v-for="(run, ridx) in filteredRuns">
                 <li :class="
@@ -33,11 +28,11 @@
                 </li>
             </template>
       </transition-group>
-        <div v-if="Math.ceil(filteredRunsTotal.length / filterLimitPerPage) > 1" class="navigation-container">
+        <div v-if="Math.ceil(filteredRunsTotal / filterLimitPerPage) > 1" class="navigation-container">
             <div class="navigation">
-                <template v-for="page in Math.ceil(filteredRunsTotal.length / filterLimitPerPage)">
-                    <div v-if="page === 1 || page === Math.ceil(filteredRunsTotal.length / filterLimitPerPage) || [currentPage-2, currentPage-1, currentPage, currentPage+1, currentPage+2].includes(page)" :class="['page',currentPage === page ? 'active' : '']" v-on:click="filterOffset = filterLimitPerPage * (page - 1); currentPage = page" :key="`page-${page}`">{{page}}</div>
-                    <div v-if="(page > 1 || page <  Math.ceil(filteredRunsTotal.length / filterLimitPerPage)) && (page === currentPage-3 && currentPage >= 5 || page === currentPage+3 && currentPage <= Math.ceil(filteredRunsTotal.length / filterLimitPerPage)-4)" class="page-offset" :key="`page-${page}`">...</div>
+                <template v-for="page in Math.ceil(filteredRunsTotal / filterLimitPerPage)">
+                    <div v-if="page === 1 || page === Math.ceil(filteredRunsTotal / filterLimitPerPage) || [currentPage-2, currentPage-1, currentPage, currentPage+1, currentPage+2].includes(page)" :class="['page',currentPage === page ? 'active' : '']" v-on:click="filterOffset = filterLimitPerPage * (page - 1); currentPage = page" :key="`page-${page}`">{{page}}</div>
+                    <div v-if="(page > 1 || page <  Math.ceil(filteredRunsTotal / filterLimitPerPage)) && (page === currentPage-3 && currentPage >= 5 || page === currentPage+3 && currentPage <= Math.ceil(filteredRunsTotal / filterLimitPerPage)-4)" class="page-offset" :key="`page-${page}`">...</div>
                 </template>
             </div>
         </div>
@@ -56,18 +51,17 @@
 // import moment from 'moment'
 import { mapRepos } from '@vuex-orm/core'
 import Run from '../store/classes/Run'
-import Tag from '../store/classes/Tag'
 import RunInfos from '../components/RunsElements/Infos.vue'
 import RunCharacter from '../components/RunsElements/Character.vue'
 import RunFloorsSlider from '../components/RunsElements/FloorsSlider.vue'
-import CustomSelect from './Tools/CustomSelect.vue'
+import RunsFilters from './RunsFilters.vue'
 export default {
     name: "Runs",
     components: {
         RunInfos,
         RunCharacter,
         RunFloorsSlider,
-        CustomSelect
+        RunsFilters
     },
     data() {
         return {
@@ -94,18 +88,12 @@ export default {
                     disable: true
                 }
             },
+            filteredRunsTotal: 0,
+            filteredRuns: [],
             currentPage: 1,
             filterLimitPerPage: 6,
             filterOffset: 0,
-            filterOrder: 'desc',
-            filterText: '',
-            filterCharacter: [],
-            filterTags: [],
-            filterCharacterVersion: '',
-            filterGameState: 0,
-            filterWiNOrDeath: '',
-            filterDateFrom: 0,
-            filterDateTo: 0
+            filterOrder: 'desc'
         }
     },
     mounted() {
@@ -113,11 +101,6 @@ export default {
         window.ipc.on('SYNC_SEND_RUNS', (response) => {
             console.log(response)
             this.runRepo.fresh(response.runs)
-        })
-        window.ipc.send('ASK_TAGS')
-        window.ipc.on('SYNC_SEND_TAGS', (response) => {
-            console.log(response)
-            this.tagRepo.fresh(response.tags)
         })
         window.ipc.on('SYNC_CREATE_RUN', (response) => {
             console.log(response)
@@ -146,23 +129,10 @@ export default {
     },
     computed: {
         ...mapRepos({
-            runRepo: Run,
-            tagRepo: Tag
+            runRepo: Run
         }),
-        allTags() {
-            return this.tagRepo.all()
-        },
-        tagsWithRuns() {
-            return this.tagRepo.where((tag) => { return this.checkTagRuns(tag) }).orderBy('value', 'desc').get()
-        },
         allRuns() {
             return this.runRepo.all()
-        },
-        filteredRunsTotal() {
-            return this.runRepo.where((run) => { return this.filter(run) }).orderBy('runUpdate', this.filterOrder).get()
-        },
-        filteredRuns() {
-            return this.filteredRunsTotal.slice(this.filterOffset, this.filterLimitPerPage + this.filterOffset)
         },
         updateRun: {
             get: function (run) {
@@ -178,41 +148,12 @@ export default {
             this.currentPage = this.currentPage === 1 ? this.currentPage : 1
             this.filterOffset = this.filterOffset === 0 ? this.filterOffset : 0
         },
-        onUpdateTagsMultiSelect(selected) {
-            this.filterTags = selected
+        onUpdateFilteredRunsTotal(total) {
+            this.filteredRunsTotal = total.length
         },
-        checkTagRuns(tag) {
-            if (tag.runs_ids.filter(runId => !!this.runRepo.where('id', runId).first()).length > 0) return tag
-        },
-        filter(run) {
-            // tags filter
-            if (this.filterTags.length > 0 && !this.filterTags.some(tag => run.tags_ids.includes(tag.id))) return
-            // text filter
-            if(this.filterText.length > 3) {
-                const textSearchValue = this.filterText.normalize('NFC').toLowerCase()
-                const characterName = run.characters[0].trueName.normalize('NFC').toLowerCase()
-                const customRunName = run.customName.normalize('NFC').toLowerCase()
-                const id = run.id.normalize('NFC').toLowerCase()
-                console.log(textSearchValue, characterName)
-                if (
-                    !characterName.includes(textSearchValue) &&
-                    !customRunName.includes(textSearchValue) &&
-                    !id.includes(textSearchValue)
-                    ) return
-            }
-            return run
+        onUpdateFilteredRuns(total) {
+            this.filteredRuns = total
         }
-        // calcBlackHeart(nb) {
-        //     if (nb === 3) return 2
-        //     if (nb === 1) return 1
-        //     let i = 0
-        //     let n = nb - 1
-        //     while (n > 1) {
-        //         n = n / 2;
-        //         i++
-        //     }
-        //     return i
-        // }
     },
 };
 </script>
@@ -226,18 +167,6 @@ export default {
 .runs-container {
     position: relative;
     padding: 0px 24px;
-}
-.filters {
-    margin-bottom: 16px;
-    text-align: left;
-    padding: 0px 12px;
-    display: flex;
-    margin-left: -8px;
-    margin-right: -8px;
-    > div {
-        margin-left: 8px;
-        margin-right: 8px;
-    }
 }
 .run {
     position: relative;
